move OptionalMailService to own service-file
