+++ tapir/core/tapir_email_base.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 
 from enum import Enum
-from typing import List, Type, TYPE_CHECKING, Dict, Tuple, Literal
-from warnings import deprecated
+from typing import List, Type, TYPE_CHECKING, Dict
 
 from django.core.mail import EmailMultiAlternatives
 from django.template import loader
@@ -215,21 +214,3 @@ class TapirEmailBase:
     @classmethod
     def register_email(cls, mail_class: Type[TapirEmailBase]):
         all_emails[mail_class.get_unique_id()] = mail_class
-
-
-class OptionalMailService:
-    @staticmethod
-    def get_optional_mail_choices() -> List[Tuple[str, str]]:
-        # this has to be a function so that choices has a callable and is refreshed whenever the form is called
-        return [
-            (mail.get_unique_id(), mail.get_name())
-            for mail in get_mail_classes(mail_option=MailOption.get_optional_options())
-        ]
-
-    @staticmethod
-    def get_mandatory_mail_choices() -> List[Tuple[str, str]]:
-        # this has to be a function so that choices has a callable and is refreshed whenever the form is called
-        return [
-            (mail.get_unique_id(), mail.get_name())
-            for mail in get_mail_classes(mail_option=MailOption.MANDATORY)
-        ]
